Revert "Merge refactor approach to next"
