+++ App.tsx
@@ -1,9 +1,11 @@
 import React from "react";
+import { ThemeProvider } from "./src/theme/themeContext";
 import Storybook from "./storybook";
 import { useFonts } from "expo-font";
 import Screen from "./src/components/Atoms/Screen/Screen";
 import AppLoading from "expo-app-loading";
-import { ThemeProvider } from "./src/theme/src/themeContext";
+import OfflineNotice from "./src/components/Molecules/OfflineNotice/OfflineNotice";
+import Button from "./src/components/Molecules/Button/Button";
 
 const App = () => {
   const [haveFontsLoaded] = useFonts({
