+++ src/components/Atoms/ActivityIndicator/ActivityIndicator.spec.js
@@ -1,6 +1,7 @@
+import { render } from "@testing-library/react-native";
 import React from "react";
 import renderer from "react-test-renderer";
-import { ThemeProvider } from "../../../theme/src/themeContext";
+import { ThemeProvider } from "../../../theme/themeContext";
 import ActivityIndicator from "./ActivityIndicator";
 
 jest.useFakeTimers();
@@ -28,5 +29,18 @@ describe("Atoms/ActivityIndicator", () => {
     });
     const tree = comp.toJSON();
     expect(tree.children.length).toBe(1);
+  });
+
+  it("overrides the color when specified", () => {
+    let comp;
+    renderer.act(() => {
+      comp = renderer.create(
+        <ThemeProvider>
+          <ActivityIndicator color="#7fff00" />
+        </ThemeProvider>
+      );
+    });
+    const tree = comp.toJSON();
+    expect(tree.children[0].props.color).toBe("#7fff00");
   });
 });
