+++ src/components/Atoms/Box/Box.spec.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Box from "./Box";
 import { render } from "@testing-library/react-native";
-import { ThemeProvider } from "../../../theme/src/themeContext";
+import { ThemeProvider } from "../../../theme/themeContext";
 
 jest.useFakeTimers();
