+++ src/components/Atoms/Box/Box.story.tsx
@@ -1,15 +1,13 @@
 import React from "react";
 import { storiesOf } from "@storybook/react-native";
 import Screen from "../Screen/Screen";
-import Box from "./Box";
+import Box from "../Box/Box";
+import Text from "../Text/Text";
 
 storiesOf("Box", module)
   .addDecorator((getStory) => <Screen>{getStory()}</Screen>)
   .add("Basic box", () => (
-    <Box
-      backgroundColor="turquoise"
-      borderRadius={20}
-      width="40%"
-      height={200}
-    ></Box>
+    <Box>
+      <Text variant="p1">This is a sample box</Text>
+    </Box>
   ));
