+++ src/components/Atoms/Screen/Screen.tsx
@@ -1,14 +1,13 @@
 import React from "react";
-import Box, { BoxProps, boxStyleFunctions } from "../Box/Box";
+import Box from "../Box/Box";
 import { SafeAreaView, StatusBar, Platform } from "react-native";
-import useStyledProps from "../../../hooks/useStyledProps";
-import { baseLightTheme } from "../../../theme/src/basetheme";
+import { BoxProps, useRestyle, boxRestyleFunctions } from "@shopify/restyle";
+import { Theme } from "../../../theme/theme";
 
-const Screen: React.FC<BoxProps<typeof baseLightTheme>> = ({
-  children,
-  ...rest
-}) => {
-  const props = useStyledProps(boxStyleFunctions, rest);
+type ScreenProps = BoxProps<Theme>;
+
+const Screen: React.FC<ScreenProps> = ({ children, ...rest }) => {
+  const props = useRestyle(boxRestyleFunctions, rest);
 
   return (
     <Box
