+++ src/components/Molecules/OfflineNotice/OfflineNotice.spec.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { render, cleanup } from "@testing-library/react-native";
 import { useNetInfo } from "@react-native-community/netinfo";
-import { ThemeProvider } from "../../../theme/src/themeContext";
+import { ThemeProvider } from "../../../theme/themeContext";
+import renderer from "react-test-renderer";
 import OfflineNotice from "./OfflineNotice";
 
 jest.useFakeTimers();
