--- src/theme/src/themeFunctions.ts
@@ -1,5 +1,0 @@
-import { BasePearlTheme } from "./types";
-
-// Enforces proper shape for theme without throwing away the user specific values
-export const createTheme = <T extends BasePearlTheme>(themeObject: T): T =>
-  themeObject;
