--- src/theme/utils/typeHelpers.ts
@@ -1,19 +1,0 @@
-import { BasePearlTheme } from "../src/types";
-
-export const getKeys = <T>(object: T) => Object.keys(object) as (keyof T)[];
-
-// export function isThemeKey<Theme extends BasePearlTheme>(
-//   theme: Theme,
-//   K: keyof Theme | undefined
-// ): boolean {
-//   if (getKeys(theme).includes(K as any)) return true;
-
-//   return false;
-// }
-
-export function isThemeKey<Theme extends BasePearlTheme>(
-  theme: Theme,
-  K: keyof Theme | undefined
-): K is keyof Theme {
-  return getKeys(theme).includes(K as keyof Theme);
-}
